Make pdf-form-filler work on node
