+++ src/entries/forms-node.ts
@@ -1,0 +1,3 @@
+export { PDFFormFiller } from '../pdf_forms/forms-node';
+export { PDFLibAdapter } from '../pdf_forms/adapters/pdf-lib';
+export { PDFJSAdapter } from '../pdf_forms/adapters/pdfjs';
